fix bug with upsertImageSampleAnalysis where it sometimes returned rowcount instead of id

+++ uk.ac.diamond.ispyb.api/src/uk/ac/diamond/ispyb/api/IspybPlateApi.java
@@ -34,7 +34,7 @@ public interface IspybPlateApi extends Closeable {
 	
 	void clearContainerError(String barcode) throws SQLException;
 
-	Integer upsertSampleImageAnalysis(SampleImageAnalysis sampleImageAnalysis);
+	Long upsertSampleImageAnalysis(SampleImageAnalysis sampleImageAnalysis);
 	
 	boolean isConnected();
 }
